requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
langgraph
faiss-cpu
sentence-transformers
PyPDF2
reportlab
duckduckgo-search
transformers
torch
